The 'path' module was split into submodules.
